MVT: Unify extent and simplify_extent validation
