+++ lib/windshaft/models/mapconfig.js
@@ -2,6 +2,7 @@ var Crypto = require('crypto');
 var semver = require('semver');
 
 var Datasource = require('./datasource');
+const _ = require('underscore');
 
 // Map configuration object
 
@@ -283,6 +284,100 @@ function create(rawConfig, datasource) {
     return new MapConfig(rawConfig, datasource);
 }
 
+
+
+/*****************************************************************************
+ * MVT
+ ****************************************************************************/
+
+const DEFAULT_EXTENT = 4096;
+const DEFAULT_SIMPLIFY_EXTENT = 256;
+
+function checkRange(number, min, max) {
+    return (!isNaN(number) && number >= min && number <= max);
+}
+
+// Checks all layers for a valid `vector_simplify_extent`
+// Makes sure all layers have the same value (or using DEFAULT_EXTENT)
+// Returns undefined if none of the layers have it declared
+function getSimplifyExtent(layers, vector_extent, callback) {
+    let undef = 0;
+    const extents = _.uniq(layers.map(layer => {
+        if (layer.options.vector_simplify_extent === undefined) {
+            undef++;
+            return layer.options.vector_extent || DEFAULT_SIMPLIFY_EXTENT;
+        }
+        return layer.options.vector_simplify_extent;
+    }));
+
+    if (extents.length > 1) {
+        return callback(new Error("Multiple simplify extent values in mapConfig (" + extents + ")"));
+    }
+
+    if (undef === layers.length) {
+        return callback(null, vector_extent);
+    }
+
+    const max_extent = vector_extent || DEFAULT_EXTENT;
+
+    // Accepted values between 1 and max_extent
+    const simplify_extent = parseInt(extents[0]);
+    if (!checkRange(simplify_extent, 1, max_extent)) {
+        return callback(new Error("Invalid vector_simplify_extent (" + simplify_extent + "). " +
+                                  "Must be between 1 and vector_extent [" + max_extent + "]"));
+    }
+
+    return callback(null, simplify_extent);
+}
+
+
+// Checks all layers for a valid `vector_extent`
+// Makes sure all layers have the same value (or using DEFAULT_EXTENT)
+// Returns undefined if none of the layers have it declared
+function getTileExtent(layers, callback) {
+    let undef = 0;
+    const layer_extents = _.uniq(layers.map(layer => {
+        if (layer.options.vector_extent === undefined) {
+            undef++;
+            return DEFAULT_EXTENT;
+        }
+        return layer.options.vector_extent;
+    }));
+
+    if (layer_extents.length > 1) {
+        return callback(new Error("Multiple extent values in mapConfig (" + layer_extents + ")"));
+    }
+
+    if (undef === layers.length) {
+        return callback(null, undefined);
+    }
+
+    // Accepted values between 1 and 2^31 -1 (2147483647)
+    const extent = parseInt(layer_extents[0]);
+    if (!checkRange(extent, 1, 2147483647)) {
+        return callback(new Error("Invalid vector_extent. Must be between 1 and 2147483647"));
+    }
+
+    return callback(null, extent);
+}
+
+MapConfig.prototype.getMVTExtents = function (callback) {
+    const layers = this.getLayers();
+
+    getTileExtent(layers, (err, extent) => {
+        if (err) {
+            return callback(err);
+        }
+
+        getSimplifyExtent(layers, extent, (err, simplify_extent) => {
+            if (err) {
+                return callback(err);
+            }
+
+            callback(null, extent || DEFAULT_EXTENT, simplify_extent || DEFAULT_SIMPLIFY_EXTENT);
+        });
+    });
+};
 
 
 module.exports = MapConfig;
