+++ lib/windshaft/renderers/mapnik/factory.js
@@ -19,8 +19,6 @@ var COLUMN_TYPE_DEFAULT = COLUMN_TYPE_GEOMETRY;
 
 var DEFAULT_TILE_SIZE = 256;
 
-const DEFAULT_EXTENT = 4096;
-const DEFAULT_SIMPLIFY_EXTENT = 256;
 
 var FORMAT_MVT = 'mvt';
 
@@ -193,21 +191,19 @@ MapnikFactory.prototype.getRenderer = function (mapConfig, format, options, call
 
     step(
         function initBuilder() {
-            var mmlBuilderOptions = {};
-            var grainstoreOptions = self._options.grainstore;
+            const mmlBuilderOptions = {};
             if (format === 'png32') {
                 mmlBuilderOptions.mapnik_tile_format = 'png';
-            } else if (format === FORMAT_MVT) {
-                // Make a copy of grainstoreOptions so we can modify it safely
-                grainstoreOptions = Object.assign({}, grainstoreOptions);
-                grainstoreOptions.datasource = Object.assign({}, grainstoreOptions.datasource);
-
-                setLayerExtent(grainstoreOptions, mmlBuilderConfig);
-                setSimplifyExtent(grainstoreOptions, mmlBuilderConfig);
             }
 
-            self._mmlStore = new grainstore.MMLStore(grainstoreOptions);
-            self._mmlStore.mml_builder(params, mmlBuilderOptions).toXML(this);
+            getMVTOptions(mapConfig, format, self._options.grainstore, (err, newOptions) => {
+                if (err) {
+                    return callback(err);
+                }
+
+                self._mmlStore = new grainstore.MMLStore(newOptions);
+                self._mmlStore.mml_builder(params, mmlBuilderOptions).toXML(this);
+            });
         },
         function loadMapnik(err, xml) {
             assert.ifError(err);
@@ -289,9 +285,7 @@ MapnikFactory.prototype.mapConfigToMMLBuilderConfig = function(mapConfig, queryR
         datasource_extend: [],
         extra_ds_opts: [],
         gcols: [],
-        'cache-features': rendererOptions.params['cache-features'],
-        layer_extents: [],
-        simplify_extents: []
+        'cache-features': rendererOptions.params['cache-features']
     };
 
     var layerFilter = rendererOptions.layer;
@@ -337,8 +331,6 @@ MapnikFactory.prototype.mapConfigToMMLBuilderConfig = function(mapConfig, queryR
         options.datasource_extend.push(mapConfig.getLayerDatasource(layerIndex));
         options.extra_ds_opts.push( extra_opt );
 
-        options.layer_extents.push(lyropt.vector_extent);
-        options.simplify_extents.push(lyropt.vector_simplify_extent);
 
         return options;
     }, options);
@@ -418,64 +410,33 @@ function isRasterColumnType(geomColumnType) {
     return geomColumnType === COLUMN_TYPE_RASTER;
 }
 
-function checkRange(number, min, max) {
-    return (!isNaN(number) && number >= min && number <= max);
-}
+function getMVTOptions(mapConfig, format, grainstoreOptions, callback) {
+    if (format !== FORMAT_MVT) {
+        return callback(null, grainstoreOptions);
+    }
 
-function setLayerExtent(grainstoreOptions, mmlOptions) {
-    const extents = _.uniq(mmlOptions.layer_extents);
-    const def = extents.indexOf(undefined);
-    if (def !== -1) {
-        if (extents.length === 1) {
-            return;
+    mapConfig.getMVTExtents((err, extent, simplify_extent) => {
+        if (err) {
+            return callback(err);
         }
-        extents[def] = DEFAULT_EXTENT;
-    }
+        // Make a copy of grainstoreOptions so we can modify it safely
+        const newOptions = Object.assign({}, grainstoreOptions);
+        newOptions.datasource = Object.assign({}, grainstoreOptions.datasource);
 
-    const extentsDefined = _.uniq(extents);
+        newOptions.datasource.vector_layer_extent = extent;
 
-    if (extentsDefined.length > 1) {
-        throw new Error("Multiple extent values in mapConfig (" + extents + ")");
-    }
+        // This function is used to set all the simplify options in Mapnik to work with MVTs
+        // so it's tied to `plugins/input/postgis/postgis_datasource.cpp`
 
-    // Accepted values between 1 and 2^31 -1 (2147483647)
-    const extent = parseInt(extentsDefined[0], 10);
-    if (!checkRange(extent, 1, 2147483647)) {
-        throw new Error("Invalid vector_extent. Must be between 1 and 2147483647");
-    }
-    grainstoreOptions.datasource.vector_layer_extent = extents[0];
-}
+        // TWKB encoding: Adapt the rounding to the simplify extent
+        newOptions.datasource.twkb_rounding_adjustment = Math.log10(extent / simplify_extent);
 
-function setSimplifyExtent(grainstoreOptions, mmlOptions) {
-    const extents = _.uniq(mmlOptions.simplify_extents);
+        // Binary encoding: Disable snapping and simplify geometries up to half pixel
+        newOptions.datasource.simplify_geometries = true;
+        newOptions.datasource.simplify_snap_ratio = 0;
+        newOptions.datasource.simplify_dp_ratio = 0.5 * extent / simplify_extent;
+        newOptions.datasource.simplify_dp_preserve = true;
 
-    const def = extents.indexOf(undefined);
-    if (def !== -1) {
-        extents[def] = grainstoreOptions.datasource.vector_layer_extent || DEFAULT_SIMPLIFY_EXTENT;
-    }
-    const extentsDefined = _.uniq(extents);
-
-    if (extentsDefined.length > 1) {
-        throw new Error("Multiple simplify extent values in mapConfig (" + extents + ")");
-    }
-
-    // Accepted values between 1 and vector_extent
-    const vector_extent = grainstoreOptions.datasource.vector_layer_extent || DEFAULT_EXTENT;
-    const simplify = parseInt(extentsDefined[0], 10);
-    if (!checkRange(simplify, 1, vector_extent)) {
-        throw new Error("Invalid vector_simplify_extent (" + simplify + "). " +
-                        "Must be between 1 and vector_extent [" + vector_extent + "]");
-    }
-
-    // This function is used to set all the simplify options in Mapnik to work with MVTs
-    // so it's tied to `plugins/input/postgis/postgis_datasource.cpp`
-
-    // TWKB encoding: Adapt the rounding to the simplify extent
-    grainstoreOptions.datasource.twkb_rounding_adjustment = Math.log10(vector_extent / simplify);
-
-    // Binary encoding: Disable snapping and simplify geometries up to half pixel
-    grainstoreOptions.datasource.simplify_geometries = true;
-    grainstoreOptions.datasource.simplify_snap_ratio = 0;
-    grainstoreOptions.datasource.simplify_dp_ratio = 0.5 * vector_extent / simplify;
-    grainstoreOptions.datasource.simplify_dp_preserve = true;
+        callback(null, newOptions);
+    });
 }
