+++ lib/windshaft/renderers/pg-mvt/factory.js
@@ -8,9 +8,6 @@ const Renderer = require('./renderer');
 const BaseAdaptor = require('../base_adaptor');
 const SubstitutionTokens = require('../../utils/substitution_tokens');
 
-const DEFAULT_EXTENT = 4096;
-const DEFAULT_SIMPLIFY_EXTENT = 256;
-
 /**
  * API: initializes the renderer, it should be called once
  *
@@ -97,73 +94,6 @@ function getLayerColumns(psql, layer) {
     });
 }
 
-function checkRange(number, min, max) {
-    return (!isNaN(number) && number >= min && number <= max);
-}
-
-// Checks all layers for a valid `vector_extent`
-// Makes sure all layers have the same value (or using DEFAULT_EXTENT)
-// Returns undefined if none of the layers have it declared
-function getTileExtent(layers, callback) {
-    let undef = 0;
-    const layer_extents = _.uniq(layers.map(layer => {
-        if (layer.options.vector_extent === undefined) {
-            undef++;
-            return DEFAULT_EXTENT;
-        }
-        return layer.options.vector_extent;
-    }));
-
-    if (layer_extents.length > 1) {
-        return callback(new Error("Multiple extent values in mapConfig (" + layer_extents + ")"));
-    }
-
-    if (undef === layers.length) {
-        return callback(null, undefined);
-    }
-
-    // Accepted values between 1 and 2^31 -1 (2147483647)
-    const extent = parseInt(layer_extents[0]);
-    if (!checkRange(extent, 1, 2147483647)) {
-        return callback(new Error("Invalid vector_extent. Must be between 1 and 2147483647"));
-    }
-
-    return callback(null, extent);
-}
-
-// Checks all layers for a valid `vector_simplify_extent`
-// Makes sure all layers have the same value (or using DEFAULT_EXTENT)
-// Returns undefined if none of the layers have it declared
-function getSimplifyExtent(layers, vector_extent, callback) {
-    let undef = 0;
-    const extents = _.uniq(layers.map(layer => {
-        if (layer.options.vector_simplify_extent === undefined) {
-            undef++;
-            return layer.options.vector_extent || DEFAULT_SIMPLIFY_EXTENT;
-        }
-        return layer.options.vector_simplify_extent;
-    }));
-
-    if (extents.length > 1) {
-        return callback(new Error("Multiple simplify extent values in mapConfig (" + extents + ")"));
-    }
-
-    if (undef === layers.length) {
-        return callback(null, vector_extent);
-    }
-
-    const max_extent = vector_extent || DEFAULT_EXTENT;
-
-    // Accepted values between 1 and max_extent
-    const simplify_extent = parseInt(extents[0]);
-    if (!checkRange(simplify_extent, 1, max_extent)) {
-        return callback(new Error("Invalid vector_simplify_extent (" + simplify_extent + "). " +
-                                  "Must be between 1 and vector_extent [" + max_extent + "]"));
-    }
-
-    return callback(null, simplify_extent);
-}
-
 module.exports = PgMvtFactory;
 const NAME = 'pg-mvt';
 const MVT_FORMAT = 'mvt';
@@ -219,25 +149,18 @@ PgMvtFactory.prototype = {
             this.options.bufferSize = mapConfig.getBufferSize(MVT_FORMAT);
         }
 
-        getTileExtent(layers, (err, extent) => {
+        mapConfig.getMVTExtents((err, extent, simplify_extent) => {
             if (err) {
                 return callback(err);
             }
 
-            this.options.vector_extent = extent || DEFAULT_EXTENT;
+            this.options.vector_extent = extent;
+            this.options.vector_simplify_extent = simplify_extent;
 
-            getSimplifyExtent(layers, extent, (err, simplify_extent) => {
-                if (err) {
-                    return callback(err);
-                }
-
-                this.options.vector_simplify_extent = simplify_extent || DEFAULT_SIMPLIFY_EXTENT;
-
-                const columnNamePromises = layers.map(layer => getLayerColumns(psql, layer));
-                Promise.all(columnNamePromises)
-                .then(() => callback(null, new Renderer(layers, psql, {}, this.options)))
-                .catch(err => callback(err));
-            });
+            const columnNamePromises = layers.map(layer => getLayerColumns(psql, layer));
+            Promise.all(columnNamePromises)
+            .then(() => callback(null, new Renderer(layers, psql, {}, this.options)))
+            .catch(err => callback(err));
         });
     }
 };
